+++ src/detection/bios/bios_bsd.c
@@ -1,12 +1,17 @@
 #include "bios.h"
 
 #include "common/settings.h"
+#include "util/smbiosHelper.h"
 
-const char* ffDetectBios(FFBiosResult* bios)
+const char* ffDetectBios(FFBiosResult* result)
 {
-    ffSettingsGetFreeBSDKenv("smbios.bios.reldate", &bios->biosDate);
-    ffSettingsGetFreeBSDKenv("smbios.bios.revision", &bios->biosRelease);
-    ffSettingsGetFreeBSDKenv("smbios.bios.vendor", &bios->biosVendor);
-    ffSettingsGetFreeBSDKenv("smbios.bios.version", &bios->biosVersion);
+    ffSettingsGetFreeBSDKenv("smbios.bios.reldate", &result->date);
+    ffCleanUpSmbiosValue(&result->date);
+    ffSettingsGetFreeBSDKenv("smbios.bios.revision", &result->biosRelease);
+    ffCleanUpSmbiosValue(&result->biosRelease);
+    ffSettingsGetFreeBSDKenv("smbios.bios.vendor", &result->vendor);
+    ffCleanUpSmbiosValue(&result->vendor);
+    ffSettingsGetFreeBSDKenv("smbios.bios.version", &result->version);
+    ffCleanUpSmbiosValue(&result->version);
     return NULL;
 }