+++ src/detection/bios/bios_windows.c
@@ -7,18 +7,18 @@ const char* ffDetectBios(FFBiosResult* bios)
     if(!ffRegOpenKeyForRead(HKEY_LOCAL_MACHINE, L"HARDWARE\\DESCRIPTION\\System\\BIOS", &hKey, NULL))
         return "ffRegOpenKeyForRead(HKEY_LOCAL_MACHINE, L\"HARDWARE\\DESCRIPTION\\System\\BIOS\", &hKey, NULL) failed";
 
-    if(!ffRegReadStrbuf(hKey, L"BIOSVersion", &bios->biosRelease, NULL))
+    if(!ffRegReadStrbuf(hKey, L"BIOSVersion", &bios->version, NULL))
         return "\"HKEY_LOCAL_MACHINE\\HARDWARE\\DESCRIPTION\\System\\BIOS\\BIOSVersion\" doesn't exist";
 
-    ffRegReadStrbuf(hKey, L"BIOSVendor", &bios->biosVendor, NULL);
-    ffRegReadStrbuf(hKey, L"BIOSReleaseDate", &bios->biosDate, NULL);
+    ffRegReadStrbuf(hKey, L"BIOSVendor", &bios->vendor, NULL);
+    ffRegReadStrbuf(hKey, L"BIOSReleaseDate", &bios->date, NULL);
 
     uint32_t major, minor;
     if(
         ffRegReadUint(hKey, L"BiosMajorRelease", &major, NULL) &&
         ffRegReadUint(hKey, L"BiosMinorRelease", &minor, NULL)
     )
-        ffStrbufAppendF(&bios->biosVersion, "%u.%u", (unsigned)major, (unsigned)minor);
+        ffStrbufAppendF(&bios->biosRelease, "%u.%u", (unsigned)major, (unsigned)minor);
 
     return NULL;
 }