+++ src/detection/board/board_bsd.c
@@ -1,10 +1,14 @@
 #include "board.h"
 #include "common/settings.h"
+#include "util/smbiosHelper.h"
 
-const char* ffDetectBoard(FFBoardResult* board)
+const char* ffDetectBoard(FFBoardResult* result)
 {
-    ffSettingsGetFreeBSDKenv("smbios.planar.product", &board->boardName);
-    ffSettingsGetFreeBSDKenv("smbios.planar.maker", &board->boardVendor);
-    ffSettingsGetFreeBSDKenv("smbios.planar.version", &board->boardVersion);
+    ffSettingsGetFreeBSDKenv("smbios.planar.product", &result->name);
+    ffCleanUpSmbiosValue(&result->name);
+    ffSettingsGetFreeBSDKenv("smbios.planar.maker", &result->vendor);
+    ffCleanUpSmbiosValue(&result->vendor);
+    ffSettingsGetFreeBSDKenv("smbios.planar.version", &result->version);
+    ffCleanUpSmbiosValue(&result->version);
     return NULL;
 }