+++ src/detection/chassis/chassis.h
@@ -7,11 +7,12 @@
 
 typedef struct FFChassisResult
 {
-    FFstrbuf chassisType;
-    FFstrbuf chassisVendor;
-    FFstrbuf chassisVersion;
+    FFstrbuf type;
+    FFstrbuf vendor;
+    FFstrbuf version;
 } FFChassisResult;
 
 const char* ffDetectChassis(FFChassisResult* result);
+const char* ffChassisTypeToString(uint32_t type);
 
 #endif