+++ src/detection/host/host_android.c
@@ -4,14 +4,7 @@
 
 const char* ffDetectHost(FFHostResult* host)
 {
-    //Family
-
-    ffStrbufInit(&host->productFamily);
     ffSettingsGetAndroidProperty("ro.product.device", &host->productFamily);
-
-    //Name
-
-    ffStrbufInit(&host->productName);
 
     ffSettingsGetAndroidProperty("ro.product.brand", &host->productName);
     if(host->productName.length > 0)
@@ -21,18 +14,9 @@ const char* ffDetectHost(FFHostResult* host)
     }
 
     ffSettingsGetAndroidProperty("ro.product.model", &host->productName);
-
     ffStrbufTrimRight(&host->productName, ' ');
 
-    //Sys vendor
-
-    ffStrbufInit(&host->sysVendor);
     ffSettingsGetAndroidProperty("ro.product.manufacturer", &host->sysVendor);
-
-    //Not implemented
-
-    ffStrbufInit(&host->productVersion);
-    ffStrbufInit(&host->productSku);
 
     return NULL;
 }