+++ src/detection/lm/lm_nosupport.c
@@ -1,6 +1,6 @@
 #include "lm.h"
 
-const char* ffDetectLM(FFLMResult* result)
+const char* ffDetectLM(FF_MAYBE_UNUSED FFLMResult* result)
 {
     return "Not supported on this platform";
 }